+++ packages/@headlessui-vue/src/components/menu/menu.ts
@@ -493,6 +493,7 @@ export let MenuItems = defineComponent({
 
 export let MenuItem = defineComponent({
   name: 'MenuItem',
+  inheritAttrs: false,
   props: {
     as: { type: [Object, String], default: 'template' },
     disabled: { type: Boolean, default: false },
@@ -571,6 +572,7 @@ export let MenuItem = defineComponent({
         role: 'menuitem',
         tabIndex: disabled === true ? undefined : -1,
         'aria-disabled': disabled === true ? true : undefined,
+        disabled: undefined, // Never forward the `disabled` prop
         onClick: handleClick,
         onFocus: handleFocus,
         onPointerenter: handleEnter,
@@ -583,7 +585,7 @@ export let MenuItem = defineComponent({
 
       return render({
         ourProps,
-        theirProps,
+        theirProps: { ...attrs, ...theirProps },
         slot,
         attrs,
         slots,